Simplification refactor
