+++ assembly/src/basic_block_builder.rs
@@ -1,6 +1,6 @@
 use alloc::{borrow::Borrow, string::ToString, sync::Arc, vec::Vec};
 
-use miden_assembly_syntax::{ast::Instruction, debuginfo::Span, diagnostics::{Report, RelatedLabel}};
+use miden_assembly_syntax::{ast::Instruction, debuginfo::Span, diagnostics::Report};
 use miden_core::{AssemblyOp, Decorator, DecoratorList, Felt, Operation, events::EventId, mast::{DecoratorId, MastNodeId}, sys_events::SystemEvent, ReducedEventID};
 
 use crate::{ProcedureContext, assembler::BodyWrapper, mast_forest_builder::MastForestBuilder};
@@ -245,9 +245,8 @@ impl BasicBlockBuilder<'_> {
     }
     
     /// Registers an EventId in the MAST Forest event table and returns the
-    /// corresponding Felt representation.
-    pub fn register_event(&mut self, event_id: EventId) -> Result<ReducedEventID, Report> {
+    /// corresponding ReducedEventID.
+    pub fn register_event(&mut self, event_id: EventId) -> ReducedEventID {
         self.mast_forest_builder.register_event(event_id)
-            .map_err(|err| Report::new(RelatedLabel::error(format!("event registration failed: {}", err))))
     }
 }
