+++ miden-vm/src/cli/mod.rs
@@ -2,7 +2,6 @@ mod bundle;
 mod compile;
 pub mod data;
 mod debug;
-mod events;
 mod prove;
 mod repl;
 mod run;
@@ -12,7 +11,6 @@ mod verify;
 pub use bundle::BundleCmd;
 pub use compile::CompileCmd;
 pub use debug::DebugCmd;
-pub use events::EventsCmd;
 pub use prove::ProveCmd;
 pub use repl::ReplCmd;
 pub use run::RunCmd;
