+++ miden-vm/src/main.rs
@@ -32,7 +32,6 @@ pub enum Actions {
     Compile(cli::CompileCmd),
     Bundle(cli::BundleCmd),
     Debug(cli::DebugCmd),
-    Events(cli::EventsCmd),
     Prove(cli::ProveCmd),
     Run(cli::RunCmd),
     Verify(cli::VerifyCmd),
@@ -48,7 +47,6 @@ impl Cli {
             Actions::Compile(compile) => compile.execute(),
             Actions::Bundle(compile) => compile.execute(),
             Actions::Debug(debug) => debug.execute(),
-            Actions::Events(events) => events.execute(),
             Actions::Prove(prove) => prove.execute(),
             Actions::Run(run) => run.execute(),
             Actions::Verify(verify) => verify.execute(),
